[M11d] Extract truncate_string/shorten_path to shared display_utils module
Fixes #548: Move duplicated truncate_string and shorten_path functions
from claude_backend.rs, codex_backend.rs, and progress.rs into a new
shared src/display_utils.rs module. The clean.rs shorten_path was left
in place as it has different semantics (home dir replacement vs last 3
path components).
